test: move fuzz test from rebuild/test to test
